+++ pubspec.yaml
@@ -98,32 +98,33 @@ dependencies:
   connectivity_plus: ^6.1.4
   flutter_link_previewer: ^3.2.2
   flutter_chat_types: ^3.6.2
-  media_kit: ^1.2.0
-  media_kit_video: ^1.3.0
-  media_kit_libs_video: ^1.0.6
+  # media_kit: ^1.2.0
+  # media_kit_video: ^1.3.0
+  # media_kit_libs_video: ^1.0.6
   visibility_detector: ^0.4.0+2
+  video_player: ^2.10.0
 
   
 
 
 dependency_overrides:
   ndk:
-    path: ../ndk/packages/ndk
-    #git:
-    #  url: https://github.com/relaystr/ndk.git
-    #  path: packages/ndk
-    #  ref: tmp-camelus-1-6-0
+    #path: ../ndk/packages/ndk
+    git:
+      url: https://github.com/relaystr/ndk.git
+      path: packages/ndk
+      ref: tmp-camelus-1-6-0
   #ndk_rust_verifier:
   #  path: ../ndk/packages/rust_verifier
     #git:
     #  url: https://github.com/relaystr/ndk.git
     #  path: packages/rust_verifier
   ndk_objectbox:
-    path: ../ndk/packages/objectbox
-    #git:
-    #   url: https://github.com/relaystr/ndk.git
-    #   path: packages/objectbox
-    #   #ref: search-usecase
+    #path: ../ndk/packages/objectbox
+    git:
+       url: https://github.com/relaystr/ndk.git
+       path: packages/objectbox
+       #ref: search-usecase
 
   #ndk_amber:
     #path: ../ndk/packages/amber
